agent/pyproject.toml
[project]
name = "uit-ai-agent-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.0",
    "crawl4ai>=0.7.4",
    "python-docx>=1.2.0",
    "fastapi>=0.118.0",
    "google-genai>=1.51.0",
    "instructor>=1.13.0",
    "langchain>=0.3.27",
    "langgraph>=0.6.8",
    "llama-index>=0.14.4",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-ollama>=0.8.0",
    "llama-index-readers-file>=0.5.4",
    "llama-index-vector-stores-chroma>=0.5.3",
    "ollama>=0.6.0",
    "openai>=1.109.1",
    "openaiapi>=0.0.2",
    "pymupdf>=1.26.6",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.37.0",
    "fastmcp>=2.13.1",
    "flagembedding>=1.3.5",
    "llama-index-tools-mcp>=0.4.3",
    "modal>=1.2.4",
    "sqlalchemy[asyncio]>=2.0.43",
    "psycopg-binary>=3.2.13",
    "alembic>=1.17.2",
    "redis>=7.1.0",
    "psycopg2-binary>=2.9.11",
    "python-multipart>=0.0.20",
    "python-jose[cryptography]>=3.5.0",
    "asyncpg>=0.31.0",
    "passlib>=1.7.4",
]

[project.scripts]
ua = "src.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
